backend/pyproject.toml
[project]
name = "shopping-ai-backend"
version = "0.1.0"
description = "Amazon Fresh Japan Shopping Assistant Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "anthropic>=0.40.0",
    "playwright>=1.49.0",
    "pyyaml>=6.0.2",
    "aiosqlite>=0.20.0",
    "sse-starlette>=2.1.3",
    "python-multipart>=0.0.12",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-pyyaml>=6.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: requires real Amazon cookies and network access",
]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
